advent_of_code v0.1.0 Mix.Tasks.Day1

Functions

detect_hq(walklog)

Detects the first position in the walklog that has been visited twice.

Examples

iex> Mix.Tasks.Day1.detect_hq([[0, 0], [1, 0], [2, 0], [2, 1], [1, 1], [1, 0], [1, -1], [0, -1], [0, 0]])
[1, 0]

manhattan_distance(x, y)

Calculates the manhattan distance for a given x and y relative to an origin at [0, 0].

Examples

iex> Mix.Tasks.Day1.manhattan_distance(1, 1)
2

iex> Mix.Tasks.Day1.manhattan_distance(14, -4)
18

run(args)

A task needs to implement run which receives a list of command line args.

Callback implementation for Mix.Task.run/1.

start_walking(movements)

Executes the given movements and returns a walklog containing a list of positions with x/y-coordinates representing the walked path.

Examples

iex> Mix.Tasks.Day1.start_walking(["R1", "R1", "R1", "R1"])
[[0, 0], [1, 0], [1, 1], [0, 1], [0, 0]]

iex> Mix.Tasks.Day1.start_walking(["R2", "L1", "L2", "L1"])
[[0, 0], [1, 0], [2, 0], [2, -1], [1, -1], [0, -1], [0, 0]]

steps(x1, y1, x2, y1)

Calculates the positions for a given movement between to coordinates.

Examples

iex> Mix.Tasks.Day1.steps(0, 0, 1, 0)
[[1, 0]]

iex> Mix.Tasks.Day1.steps(0, 0, 3, 0)
[[1, 0], [2, 0], [3, 0]]

iex> Mix.Tasks.Day1.steps(0, 0, 0, -3)
[[0, -1], [0, -2], [0, -3]]

iex> Mix.Tasks.Day1.steps(-14, 44, -14, 41)
[[-14, 43], [-14, 42], [-14, 41]]