advent_of_code v0.1.0 Mix.Tasks.Day2

Documentation for Mix.Tasks.Day2.

Functions

access_bathroom(instructions)

Hello world.

Examples

iex> Mix.Tasks.Day2.access_bathroom([[?D, ?L]])
[7]

iex> Mix.Tasks.Day2.access_bathroom([[?D, ?L], [?R, ?R, ?R, ?D, ?D]])
[7, 9]

iex> Mix.Tasks.Day2.access_bathroom([[?D, ?L], [?R, ?R, ?R, ?D, ?D], [?U, ?U, ?U, ?L, ?D]])
[7, 9, 5]

access_crazy_bathroom(instructions)

char_to_movement(char)

Maps a movement char U D L R to a movement expressed with [x, y]

Examples

iex> Mix.Tasks.Day2.char_to_movement(?U)
[0, -1]

iex> Mix.Tasks.Day2.char_to_movement(?D)
[0, 1]

iex> Mix.Tasks.Day2.char_to_movement(?L)
[-1, 0]

iex> Mix.Tasks.Day2.char_to_movement(?R)
[1, 0]

move_around_crazy_keypad(list1, list2)

Moves around a figurative keypad, respecting the physical boundaries, returning the final selected digit.

Examples

iex> Mix.Tasks.Day2.move_around_crazy_keypad([2, 0], [[-1, 0]])
[2, 0]

iex> Mix.Tasks.Day2.move_around_crazy_keypad([2, 0], [[1, 0]])
[2, 0]

iex> Mix.Tasks.Day2.move_around_crazy_keypad([2, 0], [[0, -1]])
[2, 0]

iex> Mix.Tasks.Day2.move_around_crazy_keypad([2, 0], [[0, 1]])
[2, 1]

move_around_keypad(list1, list2)

Moves around a figurative keypad, respecting the physical boundaries, returning the final selected digit.

Examples

iex> Mix.Tasks.Day2.move_around_keypad([1, 1], [[0, 1]])
[1, 2]

iex> Mix.Tasks.Day2.move_around_keypad([1, 1], [[0, 10]])
[1, 2]

iex> Mix.Tasks.Day2.move_around_keypad([1, 1], [[-10, 0]])
[0, 1]

iex> Mix.Tasks.Day2.move_around_keypad([1, 1], [[0, -1], [0, -1], [-1, 0], [-1, 0], [0, 1], [0, 1]])
[0, 2]

position_to_crazy_keypad_char(list)

Maps a x/y position to the crazy keypad char.

Examples

iex> Mix.Tasks.Day2.position_to_crazy_keypad_char([2, 0])
?1

iex> Mix.Tasks.Day2.position_to_crazy_keypad_char([1, 1])
?2

iex> Mix.Tasks.Day2.position_to_crazy_keypad_char([2, 1])
?3

iex> Mix.Tasks.Day2.position_to_crazy_keypad_char([3, 1])
?4

iex> Mix.Tasks.Day2.position_to_crazy_keypad_char([0, 2])
?5

iex> Mix.Tasks.Day2.position_to_crazy_keypad_char([1, 2])
?6

iex> Mix.Tasks.Day2.position_to_crazy_keypad_char([2, 2])
?7

iex> Mix.Tasks.Day2.position_to_crazy_keypad_char([3, 2])
?8

iex> Mix.Tasks.Day2.position_to_crazy_keypad_char([4, 2])
?9

iex> Mix.Tasks.Day2.position_to_crazy_keypad_char([1, 3])
?A

iex> Mix.Tasks.Day2.position_to_crazy_keypad_char([2, 3])
?B

iex> Mix.Tasks.Day2.position_to_crazy_keypad_char([3, 3])
?C

iex> Mix.Tasks.Day2.position_to_crazy_keypad_char([2, 4])
?D

position_to_keypad_digit(list)

Maps a x/y position to the keypad number.

Examples

iex> Mix.Tasks.Day2.position_to_keypad_digit([0, 0])
1

iex> Mix.Tasks.Day2.position_to_keypad_digit([1, 0])
2

iex> Mix.Tasks.Day2.position_to_keypad_digit([2, 0])
3

iex> Mix.Tasks.Day2.position_to_keypad_digit([0, 1])
4

iex> Mix.Tasks.Day2.position_to_keypad_digit([1, 1])
5

iex> Mix.Tasks.Day2.position_to_keypad_digit([2, 1])
6

iex> Mix.Tasks.Day2.position_to_keypad_digit([0, 2])
7

iex> Mix.Tasks.Day2.position_to_keypad_digit([1, 2])
8

iex> Mix.Tasks.Day2.position_to_keypad_digit([2, 2])
9

reduce_crazy_positions(arg1, results)

reduce_positions(arg1, results)

run(args)

A task needs to implement run which receives a list of command line args.

Callback implementation for Mix.Task.run/1.

valid_positions_on_crazy_keypad()