advent_of_code v0.1.0 Mix.Tasks.Day3

Functions

generate_vertical_triangles(triangles)

Transforms horizontal triangles to vertical triangles.

Examples

iex> Mix.Tasks.Day3.generate_vertical_triangles([[1, 2, 3], [4, 5, 6], [7, 8, 9]])
[[1, 4, 7], [2, 5, 8], [3, 6, 9]]

parse_triangle(string)

Extracts triangle from string

Examples

iex> Mix.Tasks.Day3.parse_triangle("  541  588  421")
[541, 588, 421]

iex> Mix.Tasks.Day3.parse_triangle("  530  616  422")
[530, 616, 422]

reject_invalid_triangles(triangles)

Rejects lists of invalid triangle lengths. In a valid triangle, the sum of any two sides must be larger than the remaining side. For example [5, 10, 25] is impossible, because 5 + 10 is not larger than 25.

Examples

iex> Mix.Tasks.Day3.reject_invalid_triangles([[5, 10, 25]])
[]

iex> Mix.Tasks.Day3.reject_invalid_triangles([[5, 10, 25], [2, 3, 4]])
[[2, 3, 4]]

run(args)

A task needs to implement run which receives a list of command line args.

Callback implementation for Mix.Task.run/1.