advent_of_code v0.1.0 Mix.Tasks.Day4

Functions

char_frequency(word)

Transforms horizontal triangles to vertical triangles.

Examples

iex> Mix.Tasks.Day4.char_frequency("not-a-real-room")
[{ ?o, 3 }, { ?a, 2 }, { ?r, 2 }, { ?e, 1 }, { ?l, 1 }, { ?m, 1 }, { ?n, 1 }, { ?t, 1 }]

parse(string)

Parses string, extracting its sector id, checksum and char frequency and validity

Examples

iex> Mix.Tasks.Day4.parse("aaaaa-bbb-z-y-x-123[abxyz]")
%{chars: "aaaaa-bbb-z-y-x", char_frequency: [{?a, 5}, {?b, 3}, {?x, 1}, {?y, 1}, {?z, 1}], checksum: 'abxyz', calculated_checksum: 'abxyz', sector_id: 123, valid: true}

iex> Mix.Tasks.Day4.parse("bxaxipgn-vgpst-qphzti-rdcipxcbtci-635[ipctx]")
%{chars: "bxaxipgn-vgpst-qphzti-rdcipxcbtci", char_frequency: [{?i, 4}, {?p, 4}, {?c, 3}, {?t, 3}, {?x, 3}], checksum: 'ipctx', calculated_checksum: 'ipctx', sector_id: 635, valid: true}

run(args)

A task needs to implement run which receives a list of command line args.

Callback implementation for Mix.Task.run/1.

translate_word(word, translate_by)

Parses string, extracting its sector id, checksum and char frequency and validity

Examples

iex> Mix.Tasks.Day4.translate_word("qzmt-zixmtkozy-ivhz", 343)
%{ name: "very encrypted name", id: 343 }