advent_of_code v0.1.0 Mix.Tasks.Day5

Functions

crack_password(input, password \\ "", counter \\ 1)

Cracks the password.

Examples

iex> Mix.Tasks.Day5.crack_password("abc", "18F47A3", 8605828)
"18F47A30"

crack_tougher_password(input, password \\ [nil, nil, nil, nil, nil, nil, nil, nil], counter \\ 1)

Cracks the tougher password.

Examples

iex> Mix.Tasks.Day5.crack_tougher_password("abc", ["a", nil, "a", "a", "a", "a", "a", "a"], 3231929)
"a5aaaaaa"

find_hash(payload, index \\ 1)

Finds next hash that begins with the given sequence of hex codes. May never terminate.

Examples

iex> Mix.Tasks.Day5.find_hash("abc", 3231928)
%{ hash: "00000155F8105DFF7F56EE10FA9B9ABD", id: 3231929 }

iex> Mix.Tasks.Day5.find_hash("abc", 5017307)
%{ hash: "000008F82C5B3924A1ECBEBF60344E00", id: 5017308 }

run(args)

A task needs to implement run which receives a list of command line args.

Callback implementation for Mix.Task.run/1.