advent_of_code v0.1.0 Mix.Tasks.Day8

Functions

apply_instructions(instructions)

It applies the given instructions.

Examples

iex> Mix.Tasks.Day8.apply_instructions(["rect 3x2", "rotate column x=1 by 1"])
%{ {0, 0} => 1, {0, 1} => 1, {1, 1} => 1, {1, 2} => 1, {2, 0} => 1, {2, 1} => 1}

apply_rect(map, x, y)

Creates rect of active lights with the given size.

Examples

iex> Mix.Tasks.Day8.apply_rect(%{}, 1, 1)
%{ {0, 0} => 1 }

iex> Mix.Tasks.Day8.apply_rect(%{ {0, 0} => 1}, 1, 1)
%{ {0, 0} => 1 }

iex> Mix.Tasks.Day8.apply_rect(%{}, 2, 2)
%{ {0, 0} => 1, { 0, 1 } => 1,  { 1, 0 } => 1, { 1, 1 } => 1 }

apply_rotate_column(map, col, offset)

Rotates all active lamps with a given column (or x-position) by offset. The display is 6 pixels tall (currently just hardcode) and rotating may push pixels over the “edge”.

Examples

iex> Mix.Tasks.Day8.apply_rotate_column(%{ {0, 0} => 1}, 0, 1)
%{ {0, 1} => 1 }

iex> Mix.Tasks.Day8.apply_rotate_column(%{ {0, 0} => 1}, 1, 1)
%{ {0, 0} => 1 }

iex> Mix.Tasks.Day8.apply_rotate_column(%{ {0, 0} => 1}, 0, 5)
%{ {0, 5} => 1 }

iex> Mix.Tasks.Day8.apply_rotate_column(%{ {0, 0} => 1}, 0, 6)
%{ {0, 0} => 1 }

iex> Mix.Tasks.Day8.apply_rotate_column(%{ {0, 0} => 1}, 0, -1)
%{ {0, 5} => 1 }

iex> Mix.Tasks.Day8.apply_rotate_column(%{ {0, 0} => 1, {1, 0} => 1, {2, 0} => 1}, 0, 5)
%{ {0, 5} => 1, {1, 0} => 1, {2, 0} => 1}

apply_rotate_row(map, row, offset)

Rotates all active lamps with a given row (or x-position) by offset. The display is 50 pixels wide (currently just hardcode) and rotating may push pixels over the “edge”.

Examples

iex> Mix.Tasks.Day8.apply_rotate_row(%{ {0, 0} => 1}, 0, 1)
%{ {1, 0} => 1 }

iex> Mix.Tasks.Day8.apply_rotate_row(%{ {0, 0} => 1}, 0, 49)
%{ {49, 0} => 1 }

iex> Mix.Tasks.Day8.apply_rotate_row(%{ {1, 1} => 1, {0, 0} => 1}, 1, 1)
%{ {2, 1} => 1, {0, 0} => 1 }

print(map)

run(args)

A task needs to implement run which receives a list of command line args.

Callback implementation for Mix.Task.run/1.