advent_of_code v0.1.0 Mix.Tasks.Day9

Functions

apply_marker(map)

Extracts the main three components of the currently given string.

Examples

iex> Mix.Tasks.Day9.apply_marker(%{ result: "", length: 1, times: 5, rest: "ax" })
%{ result: "aaaaa", rest: "x" }

iex> Mix.Tasks.Day9.apply_marker(%{ result: "b", length: 1, times: 5, rest: "ax" })
%{ result: "baaaaa", rest: "x" }

iex> Mix.Tasks.Day9.apply_marker(%{ result: "b", length: 2, times: 5, rest: "ax" })
%{ result: "baxaxaxaxax", rest: "" }

decompress(string)

Processes a compressed string and returns the decompressed version.

Examples

iex> Mix.Tasks.Day9.decompress("A(1x5)BC")
"ABBBBBC"

iex> Mix.Tasks.Day9.decompress("(3x3)XYZ")
"XYZXYZXYZ"

iex> Mix.Tasks.Day9.decompress("A(2x2)BCD(2x2)EFG")
"ABCBCDEFEFG"

iex> Mix.Tasks.Day9.decompress("(6x1)(1x3)A")
"(1x3)A"

iex> Mix.Tasks.Day9.decompress("X(8x2)(3x3)ABCY")
"X(3x3)ABC(3x3)ABCY"

iex> Mix.Tasks.Day9.decompress("X(10x1)1234567890")
"X1234567890"

iex> Mix.Tasks.Day9.decompress("X(1x10)1")
"X1111111111"

deep_decompress(string)

Deep decompress.

Examples

iex> Mix.Tasks.Day9.deep_decompress("X(8x2)(3x3)ABCY")
"XABCABCABCABCABCABCY"

deep_decompress_length(string)

Get deep decompress length.

Examples

iex> Mix.Tasks.Day9.deep_decompress_length("X(8x2)(3x3)ABCY")
20

iex> Mix.Tasks.Day9.deep_decompress_length("(27x12)(20x12)(13x14)(7x10)(1x12)A")
241920

iex> Mix.Tasks.Day9.deep_decompress_length("(25x3)(3x3)ABC(2x3)XY(5x2)PQRSTX(18x9)(3x2)TWO(5x7)SEVEN")
445

extract_marker(string)

Extracts the main three components of the currently given string.

Examples

iex> Mix.Tasks.Day9.extract_marker("X12(3x2)Ab4(1x1)123")
%{ result: "X12", length: 3, times: 2, rest: "Ab4(1x1)123" }

iex> Mix.Tasks.Day9.extract_marker("(3x2)Ab4(1x1)123")
%{ result: "", length: 3, times: 2, rest: "Ab4(1x1)123" }

iex> Mix.Tasks.Day9.extract_marker("(11x22)Ab4(1x1)123")
%{ result: "", length: 11, times: 22, rest: "Ab4(1x1)123" }

iex> Mix.Tasks.Day9.extract_marker("A")
%{ result: "A", length: 0, times: 0, rest: "" }

run(args)

A task needs to implement run which receives a list of command line args.

Callback implementation for Mix.Task.run/1.